Fix compilation issue for kube self-test
	[ec2-user@EC2AMAZ-N9W1Aw build]$ sudo ./credentials-fetcherd --self_test
	run diagnostic set
	krb_files_dir = /var/credentials-fetcher/krbdir
	logging_dir = /var/credentials-fetcher/logging
	unix_socket_dir = /var/credentials-fetcher/socket
	read meta data file test is successful
	<2>krb file path contains invalid characters
	read invalid metadata test is successful
	<2>directory path for kerberos tickets is not provided
	krb dir not found test is successful
	write meta data info to file test is successful
	webapp01
	/var/credentials-fetcher/krbdir/07d9d29b64b5fbc4d970/webapp01
	webapp03
	/var/credentials-fetcher/krbdir/07d9d29b64b5fbc4d970/webapp03
	parsing kube config test successful
