+++ api/tests/CMakeLists.txt
@@ -26,7 +26,7 @@ file(
         FILE_PERMISSIONS OWNER_READ OWNER_WRITE OWNER_EXECUTE GROUP_READ GROUP_EXECUTE WORLD_READ WORLD_EXECUTE
 )
 
-add_executable(gmsa_test_client "gmsa_test_client.cpp" gmsa_service_test.cpp)
+add_executable(gmsa_test_client "gmsa_test_client.cpp")
 target_link_libraries(gmsa_test_client
             cf_gmsa_service_private
             ${_PROTOBUF_LIBPROTOBUF}
